refactor: decompose wekafs.go, collect constants, sort imports
Restructures the package to match the layout the 3.0 branch uses, so subsequent
feature PRs apply against the same structure instead of colliding in a few large
files.

Splits pkg/wekafs/wekafs.go (948 lines, 31 top-level declarations) into:

  constants.go      every const declaration in the package, collected from the
                    14 files that previously held their own
  interfaces.go     VolumeType, CsiPluginMode type declarations
  apistore.go       ApiStore and its methods, ClusterApiNotFoundError
  driver.go         WekaFsDriver, its lifecycle and node-label methods,
                    initManager, vendorVersion
  leaderelection.go runWith/runWithout leader election, leader-ready file
  mounter.go        NewMounter (to be removed once mountergroup supersedes it)
  utilities.go      Die, getOwnNamespace, GetCsiPluginMode,
                    stripUnnecessaryPVFields
  wekafs_test.go    split into ascii_test.go and volume_test.go

Also moves CapacityTracker, CapacityReservation and the capacity reservation
logic out of controllerserver.go into capacitytracker.go, and regroups imports
across pkg/ and cmd/ into stdlib / third-party / local blocks.

No function body was edited. Verified by comparing the sorted set of top-level
symbols in package wekafs before and after, and by checking that all 167
declarations carrying a doc comment still carry it byte-for-byte. Two
exceptions, both deliberate:

  - a `var (...)` block mixing vendorVersion (driver concern) with
    ClusterApiNotFoundError (apistore concern) had to be split so each could
    reach its proper file; both keep their original definitions
  - the stripUnnecessaryPVFields doc comment was reworded in passing, since the
    function now serves volume health as well as capacity validation

Symbols present here but absent from the 3.0 branch - leader election, node
topology labels, stripUnnecessaryPVFields, NewMounter, GetDefaultSecrets - were
given homes rather than dropped; later feature PRs remove or supersede them.
