+++ pkg/wekafs/apiclient/apiobject.go
@@ -2,6 +2,7 @@ package apiclient
 
 import (
 	"encoding/json"
+
 	"github.com/google/uuid"
 )
