+++ pkg/wekafs/apiclient/kms.go
@@ -2,6 +2,7 @@ package apiclient
 
 import (
 	"fmt"
+
 	"golang.org/x/net/context"
 )
