+++ pkg/wekafs/apiclient/kms_test.go
@@ -2,8 +2,9 @@ package apiclient
 
 import (
 	"context"
-	"github.com/stretchr/testify/assert"
 	"testing"
+
+	"github.com/stretchr/testify/assert"
 )
 
 func TestGetKmsWhenNotDefined(t *testing.T) {
