+++ pkg/wekafs/apiclient/utils_test.go
@@ -1,9 +1,9 @@
 package apiclient
 
 import (
-	"github.com/rs/zerolog/log"
 	"testing"
 
+	"github.com/rs/zerolog/log"
 	"github.com/stretchr/testify/assert"
 )
