--- pkg/wekafs/wekafs.go
@@ -1,948 +1,0 @@
-/*
-Copyright 2017 The Kubernetes Authors.
-
-Licensed under the Apache License, Version 2.0 (the "License");
-you may not use this file except in compliance with the License.
-You may obtain a copy of the License at
-
-    http://www.apache.org/licenses/LICENSE-2.0
-
-Unless required by applicable law or agreed to in writing, software
-distributed under the License is distributed on an "AS IS" BASIS,
-WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-See the License for the specific language governing permissions and
-limitations under the License.
-*/
-
-package wekafs
-
-import (
-	"context"
-	"errors"
-	"fmt"
-	"io/fs"
-	"net"
-	"net/http"
-	"os"
-	"os/signal"
-	"strings"
-	"sync"
-	"sync/atomic"
-	"syscall"
-	"time"
-
-	"github.com/google/uuid"
-	"github.com/rs/zerolog/log"
-	"github.com/wekafs/csi-wekafs/pkg/wekafs/apiclient"
-	v1 "k8s.io/api/core/v1"
-	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
-	"k8s.io/apimachinery/pkg/runtime"
-	utilruntime "k8s.io/apimachinery/pkg/util/runtime"
-	"k8s.io/client-go/kubernetes"
-	clientgoscheme "k8s.io/client-go/kubernetes/scheme"
-	"k8s.io/client-go/rest"
-	"k8s.io/client-go/tools/clientcmd"
-	ctrl "sigs.k8s.io/controller-runtime"
-	"sigs.k8s.io/controller-runtime/pkg/cache"
-	runtimeclient "sigs.k8s.io/controller-runtime/pkg/client"
-	clog "sigs.k8s.io/controller-runtime/pkg/log"
-	"sigs.k8s.io/controller-runtime/pkg/log/zap"
-	"sigs.k8s.io/controller-runtime/pkg/manager"
-	metricsserver "sigs.k8s.io/controller-runtime/pkg/metrics/server"
-)
-
-const (
-	// LeaderStateDir is the directory where leader state files are stored
-	LeaderStateDir = "/leader-state"
-	// LeaderReadyFile signals to sidecars that this pod is the leader
-	LeaderReadyFile = "/leader-state/leader_ready"
-	// HealthProbePort is the port for the health probe endpoint
-	HealthProbePort = "8081"
-)
-
-var DefaultVolumePermissions fs.FileMode = 0750
-
-type WekaFsDriver struct {
-	name              string
-	nodeID            string
-	version           string
-	endpoint          string
-	maxVolumesPerNode int64
-	mountMode         string
-	mockMount         bool
-
-	ids            *identityServer
-	ns             *NodeServer
-	cs             *ControllerServer
-	api            *ApiStore
-	debugPath      string
-	csiMode        CsiPluginMode
-	selinuxSupport bool
-	config         *DriverConfig
-	manager        ctrl.Manager // controller-runtime manager for K8s client access
-	isLeader       atomic.Bool  // tracks current leadership state for health checks
-}
-
-type VolumeType string
-
-var (
-	vendorVersion           = "dev"
-	ClusterApiNotFoundError = errors.New("could not get API client by cluster guid")
-)
-
-// ApiStore hashmap of all APIs defined by credentials + endpoints
-type ApiStore struct {
-	sync.Mutex
-	apis          map[uint32]*apiclient.ApiClient
-	legacySecrets *map[string]string
-	config        *DriverConfig
-	Hostname      string
-}
-
-// Die used to intentionally panic and exit, while updating termination log
-func Die(exitMsg string) {
-	_ = os.WriteFile("/dev/termination-log", []byte(exitMsg), 0644)
-	panic(exitMsg)
-}
-
-// getByHash returns pointer to existing API if found by hash, or nil
-func (api *ApiStore) getByHash(key uint32) *apiclient.ApiClient {
-	if val, ok := api.apis[key]; ok {
-		return val
-	}
-	return nil
-}
-
-func (api *ApiStore) getByClusterGuid(guid uuid.UUID) (*apiclient.ApiClient, error) {
-	for _, val := range api.apis {
-		if val.ClusterGuid == guid {
-			return val, nil
-		}
-	}
-	log.Error().Str("cluster_guid", guid.String()).Msg("Could not fetch API client for cluster GUID")
-	return nil, ClusterApiNotFoundError
-}
-
-// fromSecrets returns a pointer to API by secret contents
-func (api *ApiStore) fromSecrets(ctx context.Context, secrets map[string]string, hostname string) (*apiclient.ApiClient, error) {
-	endpointsRaw := strings.TrimSpace(strings.ReplaceAll(strings.TrimSuffix(secrets["endpoints"], "\n"), "\n", ","))
-	if endpointsRaw == "" {
-		return nil, errors.New("no valid endpoints defined in secret, cannot create API client")
-	}
-	endpoints := func() []string {
-		var ret []string
-		for _, s := range strings.Split(endpointsRaw, ",") {
-			ret = append(ret, strings.TrimSpace(strings.TrimSuffix(s, "\n")))
-		}
-		return ret
-	}()
-
-	var nfsTargetIps []string
-	if _, ok := secrets["nfsTargetIps"]; ok {
-		nfsTargetIpsRaw := strings.TrimSpace(strings.ReplaceAll(strings.TrimSuffix(secrets["nfsTargetIps"], "\n"), "\n", ","))
-		nfsTargetIps = func() []string {
-			var ret []string
-			if nfsTargetIpsRaw == "" {
-				return ret
-			}
-			for _, s := range strings.Split(nfsTargetIpsRaw, ",") {
-				ret = append(ret, strings.TrimSpace(strings.TrimSuffix(s, "\n")))
-			}
-			return ret
-		}()
-	}
-
-	localContainerName, ok := secrets["localContainerName"]
-	if ok {
-		localContainerName = strings.TrimSpace(strings.TrimSuffix(localContainerName, "\n"))
-	} else {
-		localContainerName = ""
-	}
-	autoUpdateEndpoints := false
-	autoUpdateEndpointsStr, ok := secrets["autoUpdateEndpoints"]
-	if ok {
-		autoUpdateEndpoints = strings.TrimSpace(strings.TrimSuffix(autoUpdateEndpointsStr, "\n")) == "true"
-	}
-	caCertificate, ok := secrets["caCertificate"]
-	if !ok {
-		caCertificate = ""
-	}
-
-	preexistingVaultCreds := apiclient.KmsVaultCredentials{}
-
-	kmsVaultNamespaceForFilesystemEncryption, ok := secrets["kmsVaultNamespaceForFilesystemEncryption"]
-	if ok {
-		preexistingVaultCreds.Namespace = strings.TrimSpace(strings.TrimSuffix(kmsVaultNamespaceForFilesystemEncryption, "\n"))
-	}
-
-	kmsVaultKeyIdentifierForFilesystemEncryption, ok := secrets["kmsVaultKeyIdentifierForFilesystemEncryption"]
-	if ok {
-		preexistingVaultCreds.KeyIdentifier = strings.TrimSpace(strings.TrimSuffix(kmsVaultKeyIdentifierForFilesystemEncryption, "\n"))
-	}
-
-	kmsVaultRoleIdForFilesystemEncryption, ok := secrets["kmsVaultRoleIdForFilesystemEncryption"]
-	if ok {
-		preexistingVaultCreds.RoleId = strings.TrimSpace(strings.TrimSuffix(kmsVaultRoleIdForFilesystemEncryption, "\n"))
-	}
-
-	kmsVaultSecretIdForFilesystemEncryption, ok := secrets["kmsVaultSecretIdForFilesystemEncryption"]
-	if ok {
-		preexistingVaultCreds.SecretId = strings.TrimSpace(strings.TrimSuffix(kmsVaultSecretIdForFilesystemEncryption, "\n"))
-	}
-
-	credentials := apiclient.Credentials{
-		Username:            strings.TrimSpace(strings.TrimSuffix(secrets["username"], "\n")),
-		Password:            strings.TrimSuffix(secrets["password"], "\n"),
-		Organization:        strings.TrimSpace(strings.TrimSuffix(secrets["organization"], "\n")),
-		Endpoints:           endpoints,
-		HttpScheme:          strings.TrimSpace(strings.TrimSuffix(secrets["scheme"], "\n")),
-		LocalContainerName:  localContainerName,
-		AutoUpdateEndpoints: autoUpdateEndpoints,
-		CaCertificate:       caCertificate,
-		NfsTargetIPs:        nfsTargetIps,
-		KmsPreexistingCredentialsForVolumeEncryption: preexistingVaultCreds,
-	}
-	return api.fromCredentials(ctx, credentials, hostname)
-}
-
-// fromCredentials returns a pointer to API by credentials and endpoints
-// If this is a new API, it will be created and put in hashmap
-func (api *ApiStore) fromCredentials(ctx context.Context, credentials apiclient.Credentials, hostname string) (*apiclient.ApiClient, error) {
-	logger := log.Ctx(ctx)
-	logger.Trace().Str("api_client", credentials.String()).Msg("Creating new Weka API client")
-	// doing this to fetch a client hash
-	newClient, err := apiclient.NewApiClient(ctx, credentials, api.config.allowInsecureHttps, hostname)
-	if err != nil {
-		return nil, errors.New("could not create API client object from supplied params")
-	}
-	hash := newClient.Hash()
-
-	if existingApi := api.getByHash(hash); existingApi != nil {
-		logger.Trace().Str("api_client", credentials.String()).Msg("Found an existing Weka API client")
-		return existingApi, nil
-	}
-	api.Lock()
-	defer api.Unlock()
-	if api.getByHash(hash) != nil {
-		return api.getByHash(hash), nil
-	}
-	if err := newClient.Init(ctx); err != nil {
-		logger.Error().Err(err).Msg("Failed to initialize API client")
-		return nil, err
-	}
-	if !newClient.SupportsAuthenticatedMounts() && credentials.Organization != apiclient.RootOrganizationName {
-		return nil, errors.New(fmt.Sprintf(
-			"Using Organization %s is not supported on Weka cluster \"%s\".\n"+
-				"To support organization other than Root please upgrade to version %s or higher",
-			credentials.Organization, newClient.ClusterName, apiclient.MinimumSupportedWekaVersions.MountFilesystemsUsingAuthToken))
-	}
-	if (api.config.allowNfsFailback || api.config.useNfs) && !api.config.isInDevMode() {
-		newClient.NfsInterfaceGroupName = api.config.interfaceGroupName
-		newClient.NfsClientGroupName = api.config.clientGroupName
-		err := newClient.RegisterNfsClientGroup(ctx)
-		if err != nil {
-			logger.Error().Err(err).Msg("Failed to register NFS client group")
-			return nil, err
-		}
-	}
-	api.apis[hash] = newClient
-
-	return newClient, nil
-}
-
-func (api *ApiStore) GetDefaultSecrets() (*map[string]string, error) {
-	err := pathIsDirectory(LegacySecretPath)
-	if err != nil {
-		return nil, errors.New("no legacy secret exists")
-	}
-	KEYS := []string{"scheme", "endpoints", "organization", "username", "password"}
-	ret := make(map[string]string)
-	for _, k := range KEYS {
-		filePath := fmt.Sprintf("%s/%s", LegacySecretPath, k)
-		if _, err := os.Stat(filePath); os.IsNotExist(err) {
-			return nil, errors.New(fmt.Sprintf("Missing key %s in legacy secret configuration", k))
-		}
-		contents, err := os.ReadFile(filePath)
-		if err != nil {
-			return nil, errors.New(fmt.Sprintf("Could not read key %s from legacy secret configuration", k))
-		}
-		ret[k] = string(contents)
-	}
-	return &ret, nil
-}
-
-func (api *ApiStore) GetClientFromSecrets(ctx context.Context, secrets map[string]string) (*apiclient.ApiClient, error) {
-	logger := log.Ctx(ctx)
-	if len(secrets) == 0 {
-		if api.legacySecrets != nil {
-			logger.Trace().Msg("No explicit API service for request, using legacySecrets")
-			secrets = *api.legacySecrets
-		} else {
-			logger.Trace().Msg("No API service for request, switching to legacy mode")
-			return nil, nil
-		}
-	}
-	client, err := api.fromSecrets(ctx, secrets, api.Hostname)
-	if err != nil {
-		logger.Error().Err(err).Msg("Failed to initialize API client from secret, cannot proceed")
-		return nil, err
-	}
-	if client == nil {
-		logger.Trace().Msg("API service was not found for request, switching to legacy mode")
-		return nil, nil
-	}
-	logger.Trace().Msg("Successfully initialized API backend for request")
-	return client, nil
-}
-
-func NewApiStore(config *DriverConfig, hostname string) *ApiStore {
-	s := &ApiStore{
-		Mutex:    sync.Mutex{},
-		apis:     make(map[uint32]*apiclient.ApiClient),
-		config:   config,
-		Hostname: hostname,
-	}
-	secrets, err := s.GetDefaultSecrets()
-	if err != nil {
-		log.Trace().Msg("No global API secret defined")
-	} else {
-		log.Info().Msg("Initialized API with global API secret")
-		s.legacySecrets = secrets
-	}
-	return s
-}
-
-func NewWekaFsDriver(
-	driverName, nodeID, endpoint string, maxVolumesPerNode int64, version, debugPath string,
-	csiMode CsiPluginMode, selinuxSupport bool, config *DriverConfig) (*WekaFsDriver, error) {
-	if driverName == "" {
-		return nil, errors.New("no driver name provided")
-	}
-
-	if nodeID == "" {
-		return nil, errors.New("no node id provided")
-	}
-
-	if endpoint == "" {
-		return nil, errors.New("no driver endpoint provided")
-	}
-	if version != "" {
-		vendorVersion = version
-	}
-
-	log.Info().Msg(fmt.Sprintf("Driver: %v ", driverName))
-	log.Info().Msg(fmt.Sprintf("Version: %s", vendorVersion))
-
-	log.Info().Msg(fmt.Sprintf("csiMode: %s", csiMode))
-	config.Log()
-
-	return &WekaFsDriver{
-		name:              driverName,
-		nodeID:            nodeID,
-		version:           vendorVersion,
-		endpoint:          endpoint,
-		maxVolumesPerNode: maxVolumesPerNode,
-		api:               NewApiStore(config, nodeID),
-		debugPath:         debugPath,
-		csiMode:           csiMode, // either "controller", "node", "all"
-		selinuxSupport:    selinuxSupport,
-		config:            config,
-	}, nil
-}
-
-func (driver *WekaFsDriver) Run(ctx context.Context) {
-	// cleanup of stale leader file on container crash/restart
-	if driver.csiMode == CsiModeController || driver.csiMode == CsiModeAll {
-		if err := removeLeaderReadyFile(); err != nil {
-			log.Warn().Err(err).Msg("Failed to remove stale leader ready file on startup")
-		}
-	}
-
-	mounter := driver.NewMounter(ctx)
-
-	// Create servers
-	log.Info().Msg("Loading IdentityServer")
-	driver.ids = NewIdentityServer(driver.name, driver.version, driver.config)
-
-	if driver.csiMode == CsiModeController || driver.csiMode == CsiModeAll {
-		log.Info().Msg("Loading ControllerServer")
-
-		// Initialize manager with leader election
-		if err := driver.initManager(ctx, true); err != nil {
-			log.Warn().Err(err).Msg("Failed to initialize Kubernetes manager, running without leader election")
-		}
-
-		driver.cs = NewControllerServer(driver.nodeID, driver.api, mounter, driver.config, driver.manager)
-	} else {
-		driver.cs = &ControllerServer{}
-	}
-
-	if driver.csiMode == CsiModeNode || driver.csiMode == CsiModeAll {
-		// only if we manage node labels, first clean up before starting node server
-		if driver.config.manageNodeTopologyLabels {
-			log.Info().Msg("Cleaning up node stale labels")
-			driver.CleanupNodeLabels(ctx)
-		}
-
-		// In node-only mode the controller block above didn't run, so the manager (and
-		// its embedded K8s client) is not yet initialized.  Initialize it now without
-		// leader election – the node server only needs the client to read PVC/Pod
-		// annotations for per-pod mount option overrides.
-		if driver.csiMode == CsiModeNode {
-			if err := driver.initManager(ctx, false); err != nil {
-				log.Warn().Err(err).Msg("Failed to initialize Kubernetes client for node mode, per-pod mount option overrides will be unavailable")
-			}
-		}
-
-		log.Info().Msg("Loading NodeServer")
-		driver.ns = NewNodeServer(driver.nodeID, driver.maxVolumesPerNode, driver.api, mounter, driver.config)
-
-		// Read zone/region from node labels at startup so they are available
-		// when NodeGetInfo is called during registration (e.g. after CSI upgrade).
-		// This only needs to happen once — these labels are set by the cloud provider
-		// and don't change at runtime.
-		driver.readNodeTopologyLabels(ctx)
-	} else {
-		driver.ns = &NodeServer{}
-	}
-
-	s := NewNonBlockingGRPCServer(driver.csiMode)
-
-	termContext, stop := signal.NotifyContext(context.Background(), syscall.SIGTERM, syscall.SIGINT)
-	defer stop()
-
-	// Controller mode with manager: use leader election
-	// Controller mode without manager (not in K8s): run without leader election
-	// Node-only mode: run without leader election
-	if (driver.csiMode == CsiModeController || driver.csiMode == CsiModeAll) && driver.manager != nil {
-		driver.runWithLeaderElection(ctx, termContext, s)
-	} else {
-		driver.runWithoutLeaderElection(ctx, termContext, s)
-	}
-}
-
-// runWithLeaderElection runs the controller with leader election enabled.
-// Only the leader starts the gRPC server; standby pods wait for leadership.
-func (driver *WekaFsDriver) runWithLeaderElection(ctx context.Context, termContext context.Context, s *nonBlockingGRPCServer) {
-	log.Info().Msg("Running controller with leader election")
-
-	runCtx, cancelRun := context.WithCancel(ctx)
-	var shutdownOnce sync.Once
-
-	// Keep volume conditions fresh in the background so ListVolumes stays a pure cache read.
-	// Registered as an ordinary runnable, which controller-runtime gates on leadership, so exactly
-	// one replica probes the fleet and the loop stops when leadership is lost.
-	if driver.cs != nil && driver.cs.healthReconciler != nil {
-		if err := driver.manager.Add(manager.RunnableFunc(driver.cs.healthReconciler.Start)); err != nil {
-			log.Error().Err(err).Msg("Failed to register volume health reconciler, conditions will not be reported")
-		}
-	}
-
-	// Add runnable that starts gRPC server when we become leader
-	err := driver.manager.Add(manager.RunnableFunc(func(ctx context.Context) error {
-		// This only runs when we are the leader
-		log.Info().Msg("Became leader - starting gRPC server")
-
-		s.Start(driver.endpoint, driver.ids, driver.cs, driver.ns)
-
-		// Mark as leader for health checks
-		driver.isLeader.Store(true)
-
-		// Signal to sidecars that we are the leader
-		if err := createLeaderReadyFile(); err != nil {
-			log.Error().Err(err).Msg("Failed to create leader ready file")
-		}
-
-		// Wait for context cancellation (leadership lost or shutdown)
-		<-ctx.Done()
-
-		log.Info().Msg("Leadership lost or shutdown - stopping gRPC server")
-
-		// Mark as not leader for health checks
-		driver.isLeader.Store(false)
-
-		// Remove leader ready file before stopping gRPC
-		if err := removeLeaderReadyFile(); err != nil {
-			log.Error().Err(err).Msg("Failed to remove leader ready file")
-		}
-
-		s.Stop() // GracefulStop blocks until in-flight RPCs complete
-		// Lease is held until this function returns (LeaderElectionReleaseOnCancel: true)
-
-		return nil
-	}))
-	if err != nil {
-		log.Fatal().Err(err).Msg("Failed to add gRPC runnable to manager")
-	}
-
-	// Handle termination signal
-	go func() {
-		<-termContext.Done()
-		log.Info().Msg("Received termination signal")
-
-		shutdownOnce.Do(func() {
-			if (driver.csiMode == CsiModeNode || driver.csiMode == CsiModeAll) &&
-				driver.config.manageNodeTopologyLabels {
-				driver.CleanupNodeLabels(ctx)
-			}
-			cancelRun()
-		})
-	}()
-
-	// Start manager (blocks until shutdown)
-	log.Info().Msg("Starting manager - waiting for leadership")
-	if err := driver.manager.Start(runCtx); err != nil {
-		log.Error().Err(err).Msg("Manager exited with error")
-	}
-
-	s.Wait()
-	log.Info().Msg("Shutdown complete")
-}
-
-// runWithoutLeaderElection runs the driver without leader election (for node-only mode)
-func (driver *WekaFsDriver) runWithoutLeaderElection(ctx context.Context, termContext context.Context, s *nonBlockingGRPCServer) {
-	log.Info().Msg("Running without leader election (node-only mode)")
-
-	runCtx, cancelRun := context.WithCancel(ctx)
-
-	// If a manager was initialized (for K8s client access in node mode), start it so
-	// its cache syncs and GetClient() returns a working client.
-	if driver.manager != nil {
-		go func() {
-			if err := driver.manager.Start(runCtx); err != nil {
-				log.Error().Err(err).Msg("Kubernetes manager exited with error")
-			}
-		}()
-	}
-
-	go func() {
-		<-termContext.Done()
-		if (driver.csiMode == CsiModeNode || driver.csiMode == CsiModeAll) &&
-			driver.config.manageNodeTopologyLabels {
-			log.Info().Msg("Cleanup of node labels...")
-			driver.CleanupNodeLabels(ctx)
-		}
-		cancelRun()
-		s.Stop()
-		log.Info().Msg("Server stopped")
-		os.Exit(1)
-	}()
-
-	s.Start(driver.endpoint, driver.ids, driver.cs, driver.ns)
-	s.Wait()
-}
-
-// createLeaderReadyFile creates a file to signal sidecars that this pod is the leader
-func createLeaderReadyFile() error {
-	if err := os.MkdirAll(LeaderStateDir, 0o755); err != nil {
-		return fmt.Errorf("failed to create leader state directory: %w", err)
-	}
-
-	if err := os.WriteFile(LeaderReadyFile, []byte{}, 0o644); err != nil {
-		return fmt.Errorf("failed to create leader ready file: %w", err)
-	}
-
-	log.Info().Str("file", LeaderReadyFile).Msg("Created leader ready file")
-	return nil
-}
-
-// removeLeaderReadyFile removes the leader ready file when losing leadership
-func removeLeaderReadyFile() error {
-	if err := os.Remove(LeaderReadyFile); err != nil {
-		if os.IsNotExist(err) {
-			return nil
-		}
-		return fmt.Errorf("failed to remove leader ready file: %w", err)
-	}
-	log.Info().Str("file", LeaderReadyFile).Msg("Removed leader ready file (pod is not leader)")
-	return nil
-}
-
-// getOwnNamespace returns the namespace the pod is running in
-func getOwnNamespace() (string, error) {
-	// First try environment variable
-	if ns := os.Getenv("POD_NAMESPACE"); ns != "" {
-		return ns, nil
-	}
-
-	// Try to read from service account
-	nsBytes, err := os.ReadFile("/var/run/secrets/kubernetes.io/serviceaccount/namespace")
-	if err != nil {
-		return "", fmt.Errorf("failed to detect namespace: %w", err)
-	}
-	return string(nsBytes), nil
-}
-
-// initManager initializes the controller-runtime manager.
-// Pass leaderElection=true for controller mode (acquires a lease before serving).
-// Pass leaderElection=false for node mode (client-only, no lease required).
-func (d *WekaFsDriver) initManager(ctx context.Context, leaderElection bool) error {
-	logger := log.Ctx(ctx).With().Str("component", "manager-init").Logger()
-
-	// Get Kubernetes config
-	config, err := rest.InClusterConfig()
-	if err != nil {
-		if errors.Is(err, rest.ErrNotInCluster) {
-			logger.Warn().Msg("Not running in cluster, trying KUBECONFIG")
-			kubeconfig := os.Getenv("KUBECONFIG")
-			if kubeconfig == "" {
-				return fmt.Errorf("not in cluster and KUBECONFIG not set")
-			}
-			config, err = clientcmd.BuildConfigFromFlags("", kubeconfig)
-			if err != nil {
-				return fmt.Errorf("failed to build config from kubeconfig: %w", err)
-			}
-		} else {
-			return fmt.Errorf("failed to get in-cluster config: %w", err)
-		}
-	}
-
-	// Create scheme and register core v1 types
-	scheme := runtime.NewScheme()
-	utilruntime.Must(clientgoscheme.AddToScheme(scheme))
-
-	// Setup logger for controller-runtime
-	zapLogger := zap.New(zap.UseDevMode(false))
-	clog.SetLogger(zapLogger)
-
-	// When PVs are cached at all, keep only the fields the driver reads, so a large cluster's PV
-	// list stays cheap in memory.
-	cacheOpts := cache.Options{}
-	if d.config.requiresPvCaching() {
-		cacheOpts = cache.Options{
-			ByObject: map[runtimeclient.Object]cache.ByObject{
-				&v1.PersistentVolume{}: {
-					Transform: stripUnnecessaryPVFields,
-				},
-			},
-		}
-	}
-
-	mgrOpts := ctrl.Options{
-		Scheme: scheme,
-		Cache:  cacheOpts,
-		// Node mode: disable both HTTP servers to avoid port conflicts.
-		// Controller and node pods share the host network namespace (hostNetwork: true),
-		// so any port the node manager binds would block the controller manager from using it.
-		Metrics:                metricsserver.Options{BindAddress: "0"},
-		HealthProbeBindAddress: "",
-	}
-
-	if leaderElection {
-		// Controller mode: bind health probe so Kubernetes liveness probes work.
-		// Override the node-mode defaults set above.
-		healthPort := os.Getenv("HEALTH_PORT")
-		if healthPort == "" {
-			healthPort = HealthProbePort
-		}
-		mgrOpts.HealthProbeBindAddress = ":" + healthPort
-
-		// Get namespace for leader election lease
-		namespace, err := getOwnNamespace()
-		if err != nil {
-			return fmt.Errorf("failed to get namespace for leader election: %w", err)
-		}
-		mgrOpts.LeaderElection = true
-		mgrOpts.LeaderElectionNamespace = namespace
-		mgrOpts.LeaderElectionID = fmt.Sprintf("%s-controller-leader", d.name)
-		mgrOpts.LeaderElectionReleaseOnCancel = true
-	}
-
-	mgr, err := ctrl.NewManager(config, mgrOpts)
-	if err != nil {
-		return fmt.Errorf("failed to create manager: %w", err)
-	}
-
-	// ControllerGetVolume resolves a CSI volume handle back to its PersistentVolume on every
-	// health check. Index the handle so that is a cache lookup rather than a full PV scan.
-	// Registering the index starts a PV informer, so only do it where it is actually used - keyed
-	// on the csiMode that serves the controller service, which is also what gates advertising the
-	// capability, rather than on leaderElection which merely happens to correlate today.
-	servesControllerService := d.csiMode == CsiModeController || d.csiMode == CsiModeAll
-	if servesControllerService && d.config.advertiseVolumeHealthSupport {
-		if err := mgr.GetFieldIndexer().IndexField(ctx, &v1.PersistentVolume{}, pvIndexVolumeHandle,
-			func(obj runtimeclient.Object) []string {
-				pv, ok := obj.(*v1.PersistentVolume)
-				if !ok || pv.Spec.CSI == nil {
-					return nil
-				}
-				return []string{pv.Spec.CSI.VolumeHandle}
-			}); err != nil {
-			return fmt.Errorf("failed to index persistent volumes by CSI volume handle: %w", err)
-		}
-	}
-
-	if leaderElection {
-		// Parse socket path from endpoint (format: "unix:///path/to/socket")
-		socketProto, socketPath, err := parseEndpoint(d.endpoint)
-		if err != nil {
-			return fmt.Errorf("failed to parse endpoint for health check: %w", err)
-		}
-		if socketProto == "unix" {
-			socketPath = "/" + socketPath // parseEndpoint strips leading slash
-		}
-
-		// - Standby: OK (process is alive)
-		// - Leader: verify gRPC server accepts connections + Weka client running (if needed)
-		if err := mgr.AddHealthzCheck("healthz", func(r *http.Request) error {
-			if d.isLeader.Load() {
-				// Leader: verify gRPC server accepts connections
-				conn, err := net.DialTimeout(socketProto, socketPath, time.Second)
-				if err != nil {
-					return fmt.Errorf("gRPC server not accessible: %w", err)
-				}
-				_ = conn.Close()
-
-				if !d.config.useNfs && !d.config.allowNfsFailback && !d.config.isInDevMode() {
-					wekaCtx, wekaCancel := context.WithTimeout(r.Context(), d.config.healthProbeWekaTimeout)
-					defer wekaCancel()
-					if !isWekaRunning(wekaCtx) {
-						return fmt.Errorf("weka client not running or unresponsive")
-					}
-				}
-				return nil
-			}
-			// Standby: alive is enough
-			return nil
-		}); err != nil {
-			return fmt.Errorf("failed to add health check: %w", err)
-		}
-	}
-
-	d.manager = mgr
-	logger.Info().
-		Bool("leader_election", leaderElection).
-		Bool("enforce_capacity", d.config.enforceDirVolTotalCapacity).
-		Bool("advertise_volume_health_support", d.config.advertiseVolumeHealthSupport).
-		Bool("cache_persistent_volumes", d.config.requiresPvCaching()).
-		Str("leader_election_id", mgrOpts.LeaderElectionID).
-		Str("namespace", mgrOpts.LeaderElectionNamespace).
-		Msg("Kubernetes manager initialized")
-	return nil
-}
-
-// stripUnnecessaryPVFields creates a minimal PV with only fields needed for capacity validation
-func stripUnnecessaryPVFields(obj interface{}) (interface{}, error) {
-	pv, ok := obj.(*v1.PersistentVolume)
-	if !ok {
-		return obj, nil
-	}
-
-	// Build minimal PV object
-	minimal := &v1.PersistentVolume{
-		TypeMeta: pv.TypeMeta,
-		ObjectMeta: metav1.ObjectMeta{
-			Name:            pv.ObjectMeta.Name,
-			UID:             pv.ObjectMeta.UID,
-			ResourceVersion: pv.ObjectMeta.ResourceVersion,
-		},
-		Spec: v1.PersistentVolumeSpec{
-			Capacity: pv.Spec.Capacity, // Need for capacity validation
-		},
-		Status: v1.PersistentVolumeStatus{
-			Phase: pv.Status.Phase, // Need to check if Bound or Released
-		},
-	}
-
-	if pv.Spec.CSI != nil {
-		minimal.Spec.PersistentVolumeSource.CSI = &v1.CSIPersistentVolumeSource{
-			Driver:       pv.Spec.CSI.Driver,       // Need to filter by driver
-			VolumeHandle: pv.Spec.CSI.VolumeHandle, // Need to extract filesystem path
-			// ControllerGetVolume recovers the Weka API credentials from these refs, since the
-			// RPC itself carries no secrets. Refs only, the Secret contents are not cached here.
-			ControllerExpandSecretRef:  pv.Spec.CSI.ControllerExpandSecretRef,
-			ControllerPublishSecretRef: pv.Spec.CSI.ControllerPublishSecretRef,
-			NodeStageSecretRef:         pv.Spec.CSI.NodeStageSecretRef,
-			NodePublishSecretRef:       pv.Spec.CSI.NodePublishSecretRef,
-		}
-	}
-
-	return minimal, nil
-}
-
-// readNodeTopologyLabels reads the standard topology.kubernetes.io/zone and region
-// labels from the Kubernetes node object and stores them on the NodeServer.
-// Called once at startup before gRPC registration.
-func (d *WekaFsDriver) readNodeTopologyLabels(ctx context.Context) {
-	if d.ns == nil || d.manager == nil {
-		return
-	}
-	// Use the API reader (direct client) instead of the cached client because
-	// this runs before the manager is started and its informer cache is synced.
-	readCtx, cancel := context.WithTimeout(ctx, 10*time.Second)
-	defer cancel()
-	node := &v1.Node{}
-	if err := d.manager.GetAPIReader().Get(readCtx, runtimeclient.ObjectKey{Name: d.nodeID}, node); err != nil {
-		log.Warn().Err(err).Msg("Failed to get node object for reading topology labels")
-		return
-	}
-	if zone, ok := node.Labels[TopologyKeyZone]; ok {
-		d.ns.zone = zone
-	}
-	if region, ok := node.Labels[TopologyKeyRegion]; ok {
-		d.ns.region = region
-	}
-	log.Info().Str("zone", d.ns.zone).Str("region", d.ns.region).Msg("Read standard topology labels from node")
-}
-
-func (d *WekaFsDriver) SetNodeLabels(ctx context.Context) {
-	if d.config.isInDevMode() {
-		return
-	}
-
-	if d.csiMode != CsiModeNode && d.csiMode != CsiModeAll {
-		return
-	}
-	config, err := rest.InClusterConfig()
-	if err != nil {
-		log.Error().Err(err).Msg("Failed to create in-cluster config")
-		return
-	}
-
-	clientset, err := kubernetes.NewForConfig(config)
-	if err != nil {
-		log.Error().Err(err).Msg("Failed to create Kubernetes client")
-		return
-	}
-
-	node, err := clientset.CoreV1().Nodes().Get(ctx, d.nodeID, metav1.GetOptions{})
-	if err != nil {
-		log.Error().Err(err).Msg("Failed to get node object from Kubernetes")
-		return
-	}
-
-	transport := func() string {
-		if d.config.useNfs {
-			return "nfs"
-		}
-		wekaRunning := isWekaRunning(ctx)
-		if d.config.allowNfsFailback && !wekaRunning {
-			return "nfs"
-		}
-		return "wekafs"
-	}()
-
-	labelsToSet := make(map[string]string)
-	labelsToSet[TopologyKeyNode] = d.nodeID
-	labelsToSet[fmt.Sprintf(TopologyLabelNodePattern, d.name)] = d.nodeID
-	labelsToSet[fmt.Sprintf(TopologyLabelWekaLocalPattern, d.name)] = "true"
-	labelsToSet[fmt.Sprintf(TopologyLabelTransportPattern, d.name)] = transport
-	updateNeeded := false
-
-	for label, value := range labelsToSet {
-		existing, ok := node.Labels[label]
-		if !ok || existing != value {
-			log.Info().Str("label", fmt.Sprintf("%s=%s", label, value)).Str("node", node.Name).Msg("Setting label on node")
-			node.Labels[label] = value
-			updateNeeded = true
-		}
-	}
-
-	if !updateNeeded {
-		return
-	}
-
-	_, err = clientset.CoreV1().Nodes().Update(ctx, node, metav1.UpdateOptions{})
-	if err != nil {
-		log.Error().Err(err).Msg("Failed to update node labels")
-		return
-	}
-
-	log.Info().Msg("Successfully updated labels on node")
-}
-func (d *WekaFsDriver) CleanupNodeLabels(ctx context.Context) {
-	if d.config.isInDevMode() {
-		return
-	}
-	nodeLabelPatternsToRemove := []string{TopologyLabelNodePattern, TopologyLabelTransportPattern, TopologyLabelWekaLocalPattern}
-	nodeLabelsToRemove := []string{TopologyLabelTransportGlobal, TopologyLabelNodeGlobal, TopologyKeyNode}
-
-	for i, labelPattern := range nodeLabelPatternsToRemove {
-		nodeLabelPatternsToRemove[i] = fmt.Sprintf(labelPattern, d.name)
-	}
-	labelsToRemove := append(nodeLabelsToRemove, nodeLabelPatternsToRemove...)
-
-	config, err := rest.InClusterConfig()
-	if err != nil {
-		log.Error().Err(err).Msg("Failed to create in-cluster config")
-		return
-	}
-
-	clientset, err := kubernetes.NewForConfig(config)
-	if err != nil {
-		log.Error().Err(err).Msg("Failed to create Kubernetes client")
-		return
-	}
-
-	node, err := clientset.CoreV1().Nodes().Get(ctx, d.nodeID, metav1.GetOptions{})
-	if err != nil {
-		log.Error().Err(err).Msg("Failed to get node")
-		return
-	}
-
-	for _, label := range labelsToRemove {
-		delete(node.Labels, label)
-		log.Info().Str("label", label).Str("node", node.Name).Msg("Removing label from node")
-	}
-
-	_, err = clientset.CoreV1().Nodes().Update(ctx, node, metav1.UpdateOptions{})
-	if err != nil {
-		log.Error().Err(err).Msg("Failed to update node labels")
-		return
-	}
-
-	log.Info().Msg("Successfully removed labels from node")
-
-	//output, err := exec.Command("/bin/kubectl", "label", "node", d.nodeID, labelsString).Output()
-	//if err != nil {
-	//	log.Error().Err(err).Str("output", string(output)).Msg("Failed to remove labels from node")
-	//}
-}
-
-type CsiPluginMode string
-
-const (
-	VolumeTypeDirV1   VolumeType = "dir/v1"  // if specified in storage class, create directory quotas (as in legacy CSI volumes). FS name must be set in SC as well
-	VolumeTypeUnified VolumeType = "weka/v2" // no need to specify this in storageClass
-	VolumeTypeUNKNOWN VolumeType = "AMBIGUOUS_VOLUME_TYPE"
-	VolumeTypeEmpty   VolumeType = ""
-
-	LegacySecretPath = "/legacy-volume-access"
-
-	CsiModeNode       CsiPluginMode = "node"
-	CsiModeController CsiPluginMode = "controller"
-	CsiModeAll        CsiPluginMode = "all"
-)
-
-var KnownVolTypes = [...]VolumeType{VolumeTypeDirV1, VolumeTypeUnified}
-
-func GetCsiPluginMode(mode *string) CsiPluginMode {
-	ret := CsiPluginMode(*mode)
-	switch ret {
-	case CsiModeNode,
-		CsiModeController,
-		CsiModeAll:
-		return ret
-	default:
-		log.Fatal().Str("required_plugin_mode", string(ret)).Msg("Unsupported plugin mode")
-		return ""
-	}
-}
-
-func (driver *WekaFsDriver) NewMounter(ctx context.Context) AnyMounter {
-	log.Info().Msg("Configuring Mounter")
-	if driver.config.useNfs {
-		log.Warn().Msg("Enforcing NFS transport due to configuration")
-		return newNfsMounter(ctx, driver)
-	}
-	if driver.config.allowNfsFailback && !isWekaRunning(ctx) {
-		if driver.config.isInDevMode() {
-			log.Info().Msg("Not Enforcing NFS transport due to dev mode")
-		} else {
-			log.Warn().Msg("Weka Driver not found. Failing back to NFS transport")
-			return newNfsMounter(ctx, driver)
-		}
-	}
-	log.Info().Msg("Enforcing WekaFS transport")
-	return newWekafsMounter(ctx, driver)
-}
